requirements.txt
docling>=2.4.0
llama-cpp-python>=0.2.85
streamlit>=1.30.0
faiss-cpu>=1.7.0
rank_bm25>=0.2.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
